refactor(dbt): move default node functions to separate file
